+++ grails-app/init/io/xh/hoist/BootStrap.groovy
@@ -7,9 +7,24 @@
 package io.xh.hoist
 
 import grails.util.Holders
+import io.xh.hoist.admin.ConnPoolMonitoringConfig
+import io.xh.hoist.admin.MemoryMonitoringConfig
+import io.xh.hoist.alertbanner.AlertBannerConfig
 import io.xh.hoist.cluster.ClusterService
+import io.xh.hoist.config.ChangelogConfig
+import io.xh.hoist.config.IdleConfig
+import io.xh.hoist.environment.EnvPollConfig
+import io.xh.hoist.export.ExportConfig
+import io.xh.hoist.ldap.LdapConfig
+import io.xh.hoist.log.LogArchiveConfig
 import io.xh.hoist.log.LogSupport
+import io.xh.hoist.monitor.MonitorConfig
+import io.xh.hoist.telemetry.metric.MetricsConfig
+import io.xh.hoist.telemetry.trace.TraceConfig
+import io.xh.hoist.track.ActivityTrackingConfig
+import io.xh.hoist.track.ClientErrorConfig
 import io.xh.hoist.util.Utils
+import io.xh.hoist.websocket.WebSocketConfig
 
 import java.time.Instant
 import java.time.ZoneId
@@ -91,6 +106,7 @@ class BootStrap implements LogSupport {
                     maxEntriesPerMin: 1000,
                     maxRows: [default: 10000, limit: 25000, options: [1000, 5000, 10000, 25000]]
                 ],
+                typedClass: ActivityTrackingConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
                 note: 'Configures built-in Activity Tracking via TrackService.'
@@ -98,6 +114,7 @@ class BootStrap implements LogSupport {
             xhAlertBannerConfig: [
                 valueType: 'json',
                 defaultValue: [enabled: true],
+                typedClass: AlertBannerConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
                 note: 'Configures support for showing an app-wide alert banner.\n\nAdmins configure and activate alert banners from the Hoist Admin console. To generally enable this system, set "enabled" to true. The xhEnvPollConfig.interval config governs client polling for updates.'
@@ -126,6 +143,7 @@ class BootStrap implements LogSupport {
             xhChangelogConfig: [
                 valueType: 'json',
                 defaultValue: [enabled: true, excludedVersions: [], excludedCategories: [], limitToRoles: []],
+                typedClass: ChangelogConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
                 note: 'Configures built-in application changelog (release notes), with options to disable the feature entirely, exclude particular releases or categories of changes from the log, and/or only show to users with selected roles.'
@@ -133,6 +151,7 @@ class BootStrap implements LogSupport {
             xhClientErrorConfig: [
                 valueType: 'json',
                 defaultValue: [intervalMins: 2],
+                typedClass: ClientErrorConfig,
                 groupName: 'xh.io',
                 note: 'Configures handling of client error reports. Errors are queued when received and processed every [intervalMins].'
             ],
@@ -144,6 +163,7 @@ class BootStrap implements LogSupport {
                     maxSnapshots: 1440,
                     writeToLog: false
                 ],
+                typedClass: ConnPoolMonitoringConfig,
                 groupName: 'xh.io',
                 note: 'Configures built-in JDBC connection pool monitoring.'
             ],
@@ -205,6 +225,7 @@ class BootStrap implements LogSupport {
                     interval: 10,
                     onVersionChange: configService.getMap('xhAppVersionCheck', [mode: 'promptReload']).get('mode')
                 ],
+                typedClass: EnvPollConfig,
                 groupName: 'xh.io',
                 note: "Controls client calls to server to poll for version, instance changes, or auth changes. Supports the following options:\n\n" +
                     "- interval: Frequency (in seconds) with which the status of the app server should be polled. Value of -1 disables checking.\n" +
@@ -225,6 +246,7 @@ class BootStrap implements LogSupport {
                     streamingCellThreshold: 100000,
                     toastCellThreshold: 3000
                 ],
+                typedClass: ExportConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
                 note: 'Configures exporting data to Excel.'
@@ -239,6 +261,7 @@ class BootStrap implements LogSupport {
             xhIdleConfig: [
                 valueType: 'json',
                 defaultValue: [timeout: 120, appTimeouts: [:]],
+                typedClass: IdleConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
                 note: 'Governs how client application will enter "sleep mode", suspending background requests and prompting the user to reload to resume.  Timeouts are in minutes of inactivity. -1 to disable.'
@@ -250,6 +273,7 @@ class BootStrap implements LogSupport {
                     timeoutMs: 60000,
                     cacheExpireSecs: 300,
                     useMatchingRuleInChain: false,
+                    skipTlsCertVerification: false,
                     servers: [
                         [
                             host: '',
@@ -258,6 +282,7 @@ class BootStrap implements LogSupport {
                         ]
                     ]
                 ],
+                typedClass: LdapConfig,
                 groupName: 'xh.io',
                 note: 'Supports connecting to LDAP servers.'
             ],
@@ -277,6 +302,7 @@ class BootStrap implements LogSupport {
                     archiveAfterDays: 30,
                     archiveFolder: 'archive'
                 ],
+                typedClass: LogArchiveConfig,
                 groupName: 'xh.io',
                 note: 'Configures automatic cleanup and archiving of log files. Files older than "archiveAfterDays" will be moved into zipped bundles within the specified "archiveFolder".'
             ],
@@ -291,6 +317,7 @@ class BootStrap implements LogSupport {
                     maxPastInstances: 10,
                     writeToLog: true
                 ],
+                typedClass: MemoryMonitoringConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
                 note: 'Configures built-in memory usage and GC monitoring.'
@@ -306,6 +333,7 @@ class BootStrap implements LogSupport {
                     monitorTimeoutSecs: 15,
                     writeToMonitorLog: true
                 ],
+                typedClass: MonitorConfig,
                 groupName: 'xh.io',
                 note: 'Configures server-side status monitoring and notifications. Note failNotifyThreshold and warnNotifyThreshold are the number of refresh cycles a monitor will need to be in said status to trigger "alertMode".'
             ],
@@ -323,6 +351,7 @@ class BootStrap implements LogSupport {
                     prometheusConfig: [:],
                     otlpConfig: [:]
                 ],
+                typedClass: MetricsConfig,
                 groupName: 'xh.io',
                 note: 'Parameters for observable metric support'
             ],
@@ -336,6 +365,7 @@ class BootStrap implements LogSupport {
                     otlpEnabled: false,
                     otlpConfig: [:]
                 ],
+                typedClass: TraceConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
                 note: 'Parameters for distributed tracing support.'
@@ -352,6 +382,7 @@ class BootStrap implements LogSupport {
                     sendTimeLimitMs: 1000,
                     bufferSizeLimitBytes: 1000000
                 ],
+                typedClass: WebSocketConfig,
                 groupName: 'xh.io',
                 note: 'Parameters for the managed WebSocket sessions created by Hoist.'
             ]