+++ grails-app/services/io/xh/hoist/telemetry/metric/MetricsService.groovy
@@ -234,7 +234,7 @@ class MetricsService extends BaseService {
     }
 
     private MetricsConfig getConfig() {
-        new MetricsConfig(configService.getMap('xhMetricsConfig'))
+        configService.getObject(MetricsConfig)
     }